Some layout and Qt API nits and fixups
- closeEvent, showEvent, etc have been made to call super() implementation, as per
Qt API documentation recommendation.

- Gave some of the top-level windows an icon (the CashFusion icon)

- Gave the password dialog a little cute icon in the corner so it's more
of a visual cue that CashFusion is asking you for the password (and to
get the icon imprinted on people's brains)

- Renamed the "Utility..." window to "Fusions..." since for most people
that's what they probably will use it as.

- Made the "Fusions..." window also accessible from the right-click menu
on the status bar icon for power users like me that want to watch that
window as soon as fusion is enabled.