Fix: Potential workaround to missing labels
My theory as to why the labels may sometimes be missing from wallet
history is related to something I have observed on occasion: the txid
calculated by EC disagrees with the server.

So I added code to cope with that situation in the case where the server
returns a different txid on broadcast than what EC calculated.

We take the server's txid response and set a label in the wallet
for that txid as well.

While not pretty -- this workaround was added as an experiment to see if
after its addition the "disappearing labels" behavior goes away.