Removed the TxHashMismatch workaround code that was recently added
Since we know the source of this bug now, that code was unnecessary (it
lived in fusion.py after broadcasting).

We instead still look for the TxHashMismatch and log to console.  It
should never actually happen.

Also in this commit: as added paranoia we take a deepcopy of the
Transaction instance before broadcasting it for extra safety.

Also in this commit: alphabetize some of the imports at the top of
fusion.py