Nit: Made main_window get_coins return a deep copy
This is in the case where the coins it's returning are in the "pay_from"
(hard spend_from from coins tab or addresses tab).

This is because it appeared something else was modifying these coins as
the tx was being sent.  This ensures whatever takes these coins gets its
own copy of them.