Fusion password dialog: Added "virtual keyboard" support
This required some reorganization of the plugin.

Also got rid of the hidden message label for bad password since it kind
of took up lots of whitespace and also sort of interfered with the
virtual keyboard plugin.  Instead, on bad password, we just display the
message in the original message label.