Added prompting of user to start integrated Tor client
This is when the CashFusion plugin is first loaded, after 5 seconds, if
there is no Tor port autodetected and if the integrated Tor client is
not started, then we ask the user if they would like to use the
integrated Tor client.

This hopefully makes it easier on users not aware of where in the settings
to look for it.