Added "default_on" property for internal plugins
If this property appears in __init.py__ for the internal plugin, then it
will default to enabled on new installs (or installs that never
explicitly disabled the plugin).

CashFusion has this property set -- so new installs or installs that
never explicitly unloaded CashFusion will have it auto-load.

This does no harm, since auto-fusing is off by default.

Also, we set the virtual keyboard plugin to default_on -- since this is
a useful plugin for privacy/security especially if running electron cash
from an untrusted device.