Moved CashFusion settings to the Network Dialog
This closes #1784

- The CashFusion settings now lives in the Network Dialog as the only
place it lives, as a tab with an icon.
- Made the Network Dialog "Enter" key no longer close the dialog on
platforms like macOS.  It's userful to be able to edit text fields in
the network dialog and hit "Enter".
- The "UtilWin" ("Fusions..." window) now has as its parent "None", always,
regardless of who created it.
- Also gave the CashShuffle tab an icon, while I was at it.
- Added the 'on_network_dialog_jumpto' run_hook (allows to jump to
locations in the network_dialog -- CashFusion uses this.
- Renamed the SettingsDialog -> SettingsWidget
- Cleaned up a resource leak where one of the PasswordDialogs was not
being deleted.
- Ensure password dialogs die if plugin unloaded while a PasswordDialog
is up.
- Ensured "Wallet settings..." windows die immediately if the plugin is
unloaded while one of these windows is up.
- Various nits and fixups.