CashFusion: Moved the "Servers" widget into the CashFusion settings
The CashFusion settings lives in the Network dialog.  The servers widget
now also lives there.  It remains hidden if the server is not started,
but if the server is running, the widget is shown.  Since the Network
Dialog is for all things network, it was felt this is a more natural
place for the server widget for CashFusion to live.

As such it was refactored out into its own class, and a Mixin class was
created for the commonality between it and the "FusionsWindow" (e.g. it
has a timer for refreshing, etc).