CashFusion: Added status bar notification if there is connection error
Errors are:

- If Tor not found
- If cannot connect to server or protocol greeting failure

If such a thing occurs, the CashFusion icon will have a red exclamation
mark in it and a popup bubble will appear whereby the user much resolve
the error (this is similar to how CashShuffle notified users of errors).

The red exclamation mark persist until the error is resolved.

This way users are made aware when CashFusion isn't happening due to
connectivity issues, etc.