CashFusion: Bugfix + wallet settings window nit
I moved the warning in the wallet window into a separate clickable link
that pops up a help window.

I'm still not sure the warning is actually accurate, especially the way
it was reworded.  Work in progress.

The bugfix actually addresses the fact that "fraction" really was an
inverse 1.0 - fraction, rather than the percentage displayed. Thus the
UI was misleading.

This has been fixed.