CashFusion: Refactored global and per-wallet config into fusion/conf.py
This is to hopefully make it easier on future maintainers to keep track
of configuration keys and/or possibly upgrade the config keys and
per-wallet settings going forward since they all now live in 1 file.

It also hopefully will avoid undetected typos when reading/writing
settings because at least the config/settings key names appear only in 1
place in the project.