CashFusion bugfix: Coinbase was being treated improperly
- Coinbase maturity was being calculated incorrectly. (The way it was
calculated coins started off as being considered "mature" and then would
be permanently immature after 100 blocks).

- Immature coinbase coins appearing in the wallet should *not* be cause for all
fusions to cease.  Immature coinbase was being treated as if they are
unconfirmed, which with some combinations of settings would cause the
wallet to stop autofusing altogether.

The above two issues have been addressed in this commit.