CashFusion: Initial implementation of similified wallet settings
"Simple" mode features 3 modes:

 - "Normal" which strives to be a somewhat lateral shuffle
 - "Fan-out" which tends to grow the utxo set of the wallet
 - "Consolidate" which tends to shrink the utxo set of the wallet

Additionally, the old advanced mode is available as "Custom" with all
its parameters.

In the first 3 modes, the plugin decides how many threads to create
based on the number of coins, and also calculates the 'fraction'.

Note this is the initial implementation and likely this needs refining.