CashFusion: UI nit for wallet settings dialog re: Coinbase
We decided not to show the "Fuse coinbase coins" checkbox for wallets
that have no coinbase coins.  We only show this checkbox if the wallet
settings UI detects coinbase coins in the wallet.

Once latched to be shown, the checkbox never goes away.  This way
wallets that are miner's wallets get the checkbox, and everybody else is
spared the confusion.