i18n: Internal plugins were appearing untranslated, fixed.
This is because the _() decorator they were using in __init__.py wasn't
able to get the proper language string (the language is not set when
internal plugins are loaded).

So we had to write some workaround code to retranslate the pluign
"fullname" and "description" text each time before the plugin settings
dialog is shown.

Also added support for plugin "description" to be a list of _() strings
(so that the translation works properly if done later).